header rewrite traefik plugin only on openshift, refactor gateway code
